+++ packages/ai/src/generate-text/tool-set.ts
@@ -5,6 +5,10 @@ export type ToolSet = Record<
   (Tool<never, never> | Tool<any, any> | Tool<any, never> | Tool<never, any>) &
     Pick<
       Tool<any, any>,
-      'execute' | 'onInputAvailable' | 'onInputStart' | 'onInputDelta'
+      | 'execute'
+      | 'onInputAvailable'
+      | 'onInputStart'
+      | 'onInputDelta'
+      | 'needsApproval'
     >
 >;